# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yudetter SNS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /yudates:
    get:
      operationId: getTimeline
      tags: [yudates]
      summary: Get home timeline for authenticated user
      parameters:
        - name: cursor
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Timeline yudates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YudatePage"
    post:
      operationId: createYudate
      tags: [yudates]
      summary: Post a new yudate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YudateInput"
      responses:
        "201":
          description: Created yudate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Yudate"
  /yudates/{id}:
    get:
      operationId: getYudate
      tags: [yudates]
      summary: Get a single yudate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Yudate detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Yudate"
        "404":
          description: Not found
    delete:
      operationId: deleteYudate
      tags: [yudates]
      summary: Delete a yudate (author only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /yudates/{id}/like:
    post:
      operationId: likeYudate
      tags: [yudates]
      summary: Like a yudate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
    delete:
      operationId: unlikeYudate
      tags: [yudates]
      summary: Unlike a yudate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unlike result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /yudates/{id}/reyudate:
    post:
      operationId: reyudate
      tags: [yudates]
      summary: Reyudate (repost) a yudate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reyudate result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
    delete:
      operationId: unReyudate
      tags: [yudates]
      summary: Undo a reyudate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Un-reyudate result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /yudates/{id}/replies:
    get:
      operationId: getYudateReplies
      tags: [yudates]
      summary: Get replies to a yudate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Replies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YudatePage"
    post:
      operationId: replyToYudate
      tags: [yudates]
      summary: Reply to a yudate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyInput"
      responses:
        "201":
          description: Reply yudate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Yudate"

  # ── USERS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ActionResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    UserProfile:
      type: object
      properties:
        setupComplete:
          type: boolean
        isPrivate:
          type: boolean
        isBlocking:
          type: boolean
        isBlockedBy:
          type: boolean
        pinnedYudateId:
          type: ["integer", "null"]
        headerUrl:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        followerCount:
          type: integer
        followingCount:
          type: integer
        yudateCount:
          type: integer
        isFollowing:
          type: boolean
        isFollowPending:
          type: boolean
        createdAt:
          type: string
          format: date-time
        visibility:
          type: string
          enum: [public, followers]
        isSpoiler:
          type: boolean
        yudedollar:
          type: integer
        badgeType:
          type: ["string", "null"]
        consecutiveLoginDays:
          type: integer
        rankingOptIn:
          type: boolean
      required:
        - id
        - clerkId
        - username
        - displayName
        - email
        - followerCount
        - followingCount
        - yudateCount
        - isFollowing
        - isFollowPending
        - isPrivate
        - isBlocking
        - isBlockedBy
        - setupComplete
        - createdAt
        - yudedollar
        - consecutiveLoginDays
        - rankingOptIn
...
    Yudate:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        author:
          $ref: "#/components/schemas/UserProfile"
        likeCount:
          type: integer
        reyudateCount:
          type: integer
        replyCount:
          type: integer
        isLiked:
          type: boolean
        isReyudated:
          type: boolean
        quotedYudate:
          oneOf:
            - $ref: "#/components/schemas/QuotedYudate"
            - type: "null"
        replyToId:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, followers]
        isSpoiler:
          type: boolean
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/YudateReaction"
        superYudateAmount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - author
        - likeCount
        - reyudateCount
        - replyCount
        - isLiked
        - isReyudated
        - superYudateAmount
        - createdAt
    QuotedYudate:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        author:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - author
        - createdAt
    YudateInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 280
        quotedYudateId:
          type: ["integer", "null"]
      required:
        - content
    ReplyInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 280
        superYudateAmount:
          type: integer
      required:
        - content
    YudatePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Yudate"
        nextCursor:
          type: ["integer", "null"]
      required:
        - items

    # ── NOTIFICATION ─────────────────────────────────────────────────────────
...
    YudateReaction:
      type: object
      properties:
        emoji:
          type: string
        count:
          type: integer
        isReacted:
          type: boolean
      required:
        - emoji
        - count
        - isReacted
...